requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
paho-mqtt==1.6.1
requests==2.31.0
python-dotenv==1.0.0
SQLite3==3.35.0
pandas==2.1.1
scikit-learn==1.3.1
matplotlib==3.8.0
numpy==1.24.3 